+++ libyara/exec.c
@@ -883,6 +883,7 @@ int yr_execute_code(
         }
 
         result = yr_re_exec(
+          context,
           (uint8_t*) r2.re->code,
           (uint8_t*) r1.ss->c_string,
           r1.ss->length,
