+++ libyara/include/yara/utils.h
@@ -60,6 +60,14 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #endif
 
 #if defined(__GNUC__)
+#define YR_DEPRECATED_API EXTERNC __attribute__((deprecated))
+#elif defined(_MSC_VER)
+#define YR_DEPRECATED_API EXTERNC __declspec(deprecated)
+#else
+#define YR_DEPRECATED_API EXTERNC
+#endif
+
+#if defined(__GNUC__)
 #define YR_ALIGN(n) __attribute__((aligned(n)))
 #elif defined(_MSC_VER)
 #define YR_ALIGN(n) __declspec(align(n))
