+++ libyara/modules/tests.c
@@ -84,7 +84,7 @@ define_function(empty)
 
 define_function(match)
 {
-  return_integer(yr_re_match(regexp_argument(1), string_argument(2)));
+  return_integer(re_match(regexp_argument(1), string_argument(2)));
 }
